browser-research-toolkit/integrations/siri-daemon/requirements.txt
fastapi>=0.110,<1
uvicorn>=0.24,<1
pydantic>=2.0
PyYAML>=6.0.1